Fix shadowmap projection
The shadowmap projection was not a true tight fit for the scene,
sometimes objects would clip out of it if they got far enough away. This
is because I forgot to transform the scene AABB by the inverse of the
light view - instead I was transforming it by the light view.