Fix bug with physics engine
Physics was refusing to detect collisions along the z axis. That is now
fixed. What was going wrong was that the y overlap was being checked
before the z, and z was only being checked if the y overlap was not the
smallest overlap - thus, z was not being checked 90% of the time.