Remove interpolation from physics
Interpolation was added in order to smooth out the
timestepping of the physics engine, but it didn't
work that well and caused bugs on some operating systems,
so I've removed it. I plan to add some form of
interpolation back into the physics engine in the
future, but for now it works without any given a
high framerate.